DGS-24344 Move RuleMetrics to metrics package
